requirements.txt
chumpy==0.70
imageio==2.16.1
matplotlib==3.5.1
numpy==1.21.2
opencv-python==4.5.5.64
pillow==9.0.1
scikit-image==0.19.2
scikit-learn==1.0.2
scikit-video==1.1.11
scipy==1.8.0
smplx==0.1.28
trimesh==3.10.7
scikit-learn
lmdb
ninja
requests
tqdm
configargparse
munch
imageio_ffmpeg
pytorch_fid